requirements.txt
Django==5.2.10
openpyxl==3.1.5
django-unfold==0.89.0